First reactive flow

+++ sea-shell-soap-wiremock/sea-shell-cli-accounts/src/main/java/org/jesperancinha/shell/client/accounts/SeaShellsWSDLAccounts.java
@@ -25,6 +25,6 @@ public interface SeaShellsWSDLAccounts {
     public Account accounts(
 
             @WebParam(name = "accountId", targetNamespace = "")
-                    int accountId
+                    long accountId
     );
 }
